+++ layers/Schema/packages/post-categories/src/FieldResolvers/PostCategoryListFieldResolver.php
@@ -20,8 +20,8 @@ class PostCategoryListFieldResolver extends AbstractPostFieldResolver
         $descriptions = [
             'posts' => $this->translationAPI->__('Posts which contain this category', 'post-categories'),
             'postCount' => $this->translationAPI->__('Number of posts which contain this category', 'post-categories'),
-            'unrestrictedPosts' => $this->translationAPI->__('[Unrestricted] Posts which contain this category', 'post-categories'),
-            'unrestrictedPostCount' => $this->translationAPI->__('[Unrestricted] Number of posts which contain this category', 'post-categories'),
+            'postsForAdmin' => $this->translationAPI->__('[Unrestricted] Posts which contain this category', 'post-categories'),
+            'postCountForAdmin' => $this->translationAPI->__('[Unrestricted] Number of posts which contain this category', 'post-categories'),
         ];
         return $descriptions[$fieldName] ?? parent::getSchemaFieldDescription($typeResolver, $fieldName);
     }
@@ -42,8 +42,8 @@ class PostCategoryListFieldResolver extends AbstractPostFieldResolver
         switch ($fieldName) {
             case 'posts':
             case 'postCount':
-            case 'unrestrictedPosts':
-            case 'unrestrictedPostCount':
+            case 'postsForAdmin':
+            case 'postCountForAdmin':
                 $query['category-ids'] = [$typeResolver->getID($category)];
                 break;
         }